feat: added excpetion handler validation

+++ src/users/commands/handlers/delete-user.handler.ts
@@ -1,5 +1,5 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
-import { Inject, Injectable } from '@nestjs/common';
+import { Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { Model } from 'mongoose';
 import { User } from 'src/users/interfaces/user.interface';
 import { DeleteUserCommand } from '../impl/delete-user.command';
@@ -10,12 +10,12 @@ export class DeleteUserHandler implements ICommandHandler<DeleteUserCommand> {
   constructor(@Inject('USER_MODEL') private readonly userModel: Model<User>) {}
 
   async execute(command: DeleteUserCommand) {
-    // find user by id
-    // if (!user) {
-    //   throw new Error('User not found');
-    // }
-    console.log('command', command);
-    const ok = await this.userModel.deleteOne({ _id: command.id }).exec();
-    console.log('ok', ok);
+    const user = await this.userModel.findById(command.id);
+    // find user by email
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    await this.userModel.deleteOne({ _id: command.id }).exec();
   }
 }
